Add an example of context menu for work packages
To add a context menu entry for single work packages, the frontend part
of the plugin must register it's context menu as outlined on `main.ts`

The menu entry depends on the `link` property that is defined on the backend
on `engine.rb`

On the backend part we can skip creating a link property in specific
applications. Eg. User not being allowed to see the context menu.

It's currently not possible to add context menu entries for multiple
selection of work packages.